lesson2/seminar/step5_homework_model_validation_and_metrics/pyproject.toml
[project]
name = "step5-homework-model-validation-and-metrics"
version = "0.1.0"
description = ""
authors = [
    {name = "SIVVIT",email = "47886214+SIVVIT@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas (>=2.2,<3.0)",
    "scikit-learn (>=1.5,<2.0)",
    "dvc (>=3.50,<4.0)",
    "great-expectations (>=0.18,<1.0)",
    "pyyaml (>=6.0,<7.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pre-commit (>=4.3.0,<5.0.0)",
    "black (>=25.9.0,<26.0.0)",
    "ruff (>=0.14.1,<0.15.0)",
    "isort (>=7.0.0,<8.0.0)"
]
